fix(tsc-wrapped): decouple bundle index host from tsickle dependency (#18999)
PR Close #18999
